+++ src/components/lists/scheduling/item/index.tsx
@@ -1,81 +1,13 @@
-import { View, Text, Linking } from "react-native";
+import { View, Text } from "react-native";
 import { SchedulingItemStyles as styles } from "./styles";
 import { Schedule } from "../../../../models/schedule.model";
 import {
   MaterialIcons as Icon,
   MaterialCommunityIcons as Icons,
 } from "@expo/vector-icons";
+import { adjustPhone, adjustTime, onPressPhoneCall, onPressWhatsApp } from "../../../../services/phone.service";
 
 export function SchedulingItem({ data }: { data: Schedule }) {
-  const adjustPhone = (phone: string) => {
-    phone = phone.trim().replace(/\s/g, '').replace("-", "");
-    phone = phone.startsWith("0") ? phone.slice(1) : phone;
-    let ddd = "";
-    let _phone = "";
-    let phone_ = "";
-    let pattern = "";
-    if (phone.length == 11) {
-      ddd = phone.slice(0, 2);
-      _phone = phone.slice(2, 7);
-      phone_ = phone.slice(7, 11);
-      pattern = `(${ddd}) ${_phone}-${phone_}`;
-    } else if (phone.length == 10) {
-      ddd = phone.slice(0, 2);
-      _phone = phone.slice(2, 6);
-      phone_ = phone.slice(6, 10);
-      pattern = `(${ddd}) 9${_phone}-${phone_}`;
-    } else {
-      return phone
-    }
-    return pattern;
-  };
-
-  const adjustTime = (time: string) => {
-    time = time.trim().replace(/\s/g, '').replace(":", "");
-    if (time.length == 1) {
-      time = `0${time}:00`;
-    } else if (time.length == 2) {
-      time = `${time}:00`;
-    } else {
-      let hour = time.slice(0, 2);
-      let minute = time.slice(2, 4);
-      time = `${hour}:${minute}`;
-    }
-    return time;
-  };
-
-  const onPressWhatsApp = (phone: string) => {
-    phone = phone.trim().replace(/\s/g, '').replace("-", "");
-    if (phone.startsWith("0")) {
-      phone = phone.slice(1);
-    }
-    phone = `55${phone}`;
-    let text = `Olá *${data.nomeCliente}*, você tem uma visita agendada para hoje, às *${data.horaInicial}*, podemos confirmar? Senão, podemos remarcar?`;
-    const link = async () => {
-      await Linking.canOpenURL(`whatsapp://send?text=${text}`).then((suppoted) => {
-        if (suppoted) {
-          return Linking.openURL(`whatsapp://send?phone=${phone}&text=${text}`);
-        } else {
-          return Linking.openURL(
-            `https://api.whatsapp.com/send?phone=${phone}&text=${text}`
-          );
-        }
-      });
-    }
-    link().catch(error => console.error(error))
-  };
-
-  const onPressPhoneCall = (phone: string) => {
-    phone = phone.trim().replace(/\s/g, '').replace("-", "");
-    if (phone.startsWith("0")) {
-      phone = phone.slice(1);
-    }
-    const link = async () => {
-      await Linking.openURL(`tel:${phone}`);
-    }
-    link().catch(error => console.error(error))
-  };
-
   return (
     <View style={styles.container}>
       <View style={styles.top}>
@@ -125,7 +57,7 @@ export function SchedulingItem({ data }: { data: Schedule }) {
                 color="green"
                 size={24}
                 onPress={() => {
-                  onPressWhatsApp(data.telefone);
+                  onPressWhatsApp(data.telefone, data.nomeCliente, data.horaInicial);
                 }} />
             </>
             : undefined}
